+++ test/libsolidity/syntaxTests/constants/cyclic_dependency_1.sol
@@ -1,0 +1,5 @@
+contract C {
+    uint constant a = a;
+}
+// ----
+// TypeError: The value of the constant a has a cyclic dependency via a.