+++ test/libsolidity/syntaxTests/structs/recursion/parallel_structs.sol
@@ -1,0 +1,15 @@
+pragma experimental ABIEncoderV2;
+
+contract TestContract
+{
+    struct SubStruct {
+        uint256 id;
+    }
+    struct TestStruct {
+        SubStruct subStruct1;
+        SubStruct subStruct2;
+    }
+    function addTestStruct(TestStruct) public pure {}
+}
+// ----
+// Warning: Experimental features are turned on. Do not use experimental features on live deployments.